create dedicated UserAvatar component
